helm-version-checker/requirements.txt
requests==2.32.5
pyyaml==6.0.3
